requirements.txt
regolo
python-dotenv